requirements.txt
openai
pyyaml
feedparser
requests
beautifulsoup4
